refactor(backend): Moved tty backend code into tty/ folder
